refs/heads/ui-test-automation Extract super class
